move and document import/export functions
