+++ extension/data/tbcore.js
@@ -872,57 +872,6 @@ export async function getConfig (sub) {
     return resp;
 }
 
-// TODO: Move this function to tbmodule, the only place it's ever used
-export async function exportSettings (subreddit, callback) {
-    const settingsObject = await TBStorage.getSettings();
-
-    // Transform from the normal setting storage format to the backup format
-    const backupObject = Object.fromEntries(
-        Object.entries(settingsObject)
-            // In extension storage, keys are `Toolbox.Module.settingName` - in the
-            // backup format however they're just `Module.settingName`. Let's convert
-            .map(([key, value]) => [key.replace('Toolbox.', ''), value])
-            // don't backup the setting registry
-            // TODO: wait, what is this??
-            .filter(([key]) => key !== 'Storage.setting')
-            // DO NOT, EVER save null (or undefined, but we shouldn't ever get
-            // that)
-            .filter(([_key, value]) => value != null),
-    );
-
-    TBApi.postToWiki('tbsettings', subreddit, backupObject, 'exportSettings', true, false).then(callback);
-}
-
-// TODO: Move this function to tbmodule, the only place it's ever used
-export async function importSettings (subreddit) {
-    const resp = await TBApi.readFromWiki(subreddit, 'tbsettings', true);
-    if (!resp || resp === TBApi.WIKI_PAGE_UNKNOWN || resp === TBApi.NO_WIKI_PAGE) {
-        logger.log('Error loading wiki page');
-        return;
-    }
-    TBStorage.purifyObject(resp);
-
-    if (resp['Utils.lastversion'] < 300) {
-        logger.log('Cannot import from a toolbox version under 3.0');
-        return;
-    }
-
-    const doNotImport = [
-        'oldreddit.enabled',
-    ];
-
-    const newSettings = Object.fromEntries(
-        Object.entries(resp)
-            // Exclude certain settings whose values shouldn't be changed
-            .filter(([key]) => !(doNotImport.includes(key)))
-            // In backups, keys are `Module.settingName` - in extension storage
-            // they're `Toolbox.Module.settingName`
-            .map(([key, value]) => [`Toolbox.${key}`, value]),
-    );
-
-    await TBStorage.writeSettings(newSettings);
-}
-
 // Misc. functions
 
 export function addToSiteTable (URL, callback) {
